Rename Characters to Entities
